\name{colcenter}
\alias{colcenter}
\title{column centering of the data matrix
}
\description{This function centerizes each column of the data matrix
}
\usage{
colcenter(X)
}

\arguments{
  \item{X}{input data matrix
}
}
\value{
return the data matrix with each column centered
}
\references{Computation of ancestry scores with mixed families and unrelated individuals. Yi-Hui Zhou, J.S. Marron, Fred Wright, arXiv:1606.08416.
}
\author{Yi-Hui Zhou
}

