\name{rowscale}
\alias{rowscale}
\title{
Scale each row of a matrix
}
\description{
This function scales the input matrix so that each row mean is 0 and each row (sample) variance is 1.
}
\usage{
rowscale(X)
}
\arguments{
\item{X}{input data matrix}
}
\value{
Output the row-scaled matrix.
}
\references{
Computation of ancestry scores with mixed families and unrelated individuals.  arXiv:1606.08416.
}
\author{
Yi-Hui ZHou, Fred A. Wright
}
